# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * D&D Text Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface CreatePlayerRequest {
  username: string;
  password: string;
}

export interface Player {
  id: number;
  username: string;
  createdAt: string;
}

export interface ValidateCharacterRequest {
  name: string;
  race: string;
  class: string;
  backstory?: string;
}

export interface ValidateCharacterResponse {
  valid: boolean;
  message: string;
  suggestion?: string;
}

export interface CreateCharacterRequest {
  name: string;
  race: string;
  class: string;
  backstory?: string;
}

export interface Character {
  id: number;
  playerId: number;
  name: string;
  race: string;
  class: string;
  backstory?: string;
  level: number;
  hp: number;
  maxHp: number;
  xp: number;
  isDead?: boolean;
  createdAt: string;
}

export interface InventoryItem {
  id: number;
  characterId: number;
  campaignId: number;
  name: string;
  description?: string;
  type: string;
  quantity: number;
  createdAt: string;
}

export interface AddInventoryItemRequest {
  name: string;
  description?: string;
  type: string;
  quantity: number;
}

export interface Achievement {
  id: number;
  playerId: number;
  title: string;
  description: string;
  icon: string;
  unlockedAt: string;
}

export type CreateCampaignRequestDmType =
  (typeof CreateCampaignRequestDmType)[keyof typeof CreateCampaignRequestDmType];

export const CreateCampaignRequestDmType = {
  ai: "ai",
  player: "player",
} as const;

export interface CreateCampaignRequest {
  title: string;
  description: string;
  setting: string;
  isPublic: boolean;
  creatorId: number;
  inviteCode?: string;
  dmType?: CreateCampaignRequestDmType;
  humanDmId?: number;
}

export type CampaignDmType =
  (typeof CampaignDmType)[keyof typeof CampaignDmType];

export const CampaignDmType = {
  ai: "ai",
  player: "player",
} as const;

export interface Campaign {
  id: number;
  title: string;
  description: string;
  setting: string;
  isPublic: boolean;
  creatorId: number;
  inviteCode?: string;
  dmType: CampaignDmType;
  humanDmId?: number;
  createdAt: string;
}

export interface JoinCampaignRequest {
  playerId: number;
  characterId: number;
  inviteCode?: string;
}

export interface GameSession {
  id: number;
  campaignId: number;
  playerId: number;
  characterId: number;
  status: string;
  createdAt: string;
}

export interface PostDiscussionRequest {
  playerId: number;
  username: string;
  message: string;
}

export interface DiscussionMessage {
  id: number;
  campaignId: number;
  playerId: number;
  username: string;
  message: string;
  createdAt: string;
}

export interface PartyMember {
  playerId: number;
  username: string;
  characterName: string;
  race: string;
  class: string;
  level: number;
  hp: number;
  maxHp: number;
  xp: number;
  isDead?: boolean;
}

export interface DmInjectRequest {
  creatorId: number;
  event: string;
}

/**
 * When present, the server performs a trusted dice roll using this notation and ignores the action text.
 */
export type GameActionRequestDiceRoll = {
  /** Dice notation like "1d20" or "2d6" */
  notation: string;
};

export interface GameActionRequest {
  action: string;
  characterId?: number;
  playerId?: number;
  /** When true, the message is human-DM narration and is stored directly without AI. */
  isDmNarration?: boolean;
  /** When present, the server performs a trusted dice roll using this notation and ignores the action text. */
  diceRoll?: GameActionRequestDiceRoll;
}

export interface GameActionResponse {
  narrative: string;
  sessionId: number;
  xpGained?: number;
  hpChange?: number;
  diceRequest?: string;
  isDead?: boolean;
  /** True when the campaign uses a human DM and the DM has not yet responded. */
  awaitingDm?: boolean;
  newAchievements?: Achievement[];
  newItems?: InventoryItem[];
  newLocation?: string;
}

export interface GameMessage {
  id: number;
  sessionId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface Npc {
  id: number;
  sessionId: number;
  name: string;
  description: string;
  disposition: string;
  firstMet?: string;
}

export interface JournalEntry {
  id: number;
  sessionId: number;
  summary: string;
  createdAt: string;
}

export interface WorldMap {
  sessionId: number;
  locations: string[];
  currentLocation?: string;
  ascii?: string;
}

export interface Combatant {
  name: string;
  hp: number;
  maxHp: number;
  initiative: number;
  isPlayer: boolean;
}

export interface CombatState {
  sessionId: number;
  active: boolean;
  round: number;
  combatants: Combatant[];
}

export interface UpdateCombatRequest {
  active: boolean;
  round: number;
  combatants: Combatant[];
}

export interface Notice {
  id: number;
  playerId: number;
  username: string;
  content: string;
  campaignTitle?: string;
  createdAt: string;
}

export interface PostNoticeRequest {
  playerId: number;
  username: string;
  content: string;
  campaignTitle?: string;
}

export interface UpdateMemberStatsRequest {
  hp: number;
  maxHp: number;
  level: number;
  xp: number;
}

export interface DmDiceRequest {
  targetCharacterId?: number;
  targetPlayerId?: number;
  notation: string;
}

export type GetCampaignsParams = {
  playerId?: number;
};