# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: D&D Text Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns/{campaignId}/party:
    get:
      operationId: getCampaignParty
      tags: [campaigns]
      summary: Get all active players in a campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Party members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartyMember"
  /campaigns/{campaignId}/members/{playerId}/stats:
    put:
      operationId: updateMemberStats
      tags: [campaigns]
      summary: Update stats for a campaign member (DM only)
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberStatsRequest"
      responses:
        "200":
          description: Member stats updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /campaigns/{campaignId}/dm-inject:
    post:
      operationId: dmInjectEvent
      tags: [campaigns]
      summary: Campaign creator injects a DM event into all sessions
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DmInjectRequest"
      responses:
        "200":
          description: Event injected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /sessions/{sessionId}/npcs:
    get:
      operationId: getSessionNpcs
      tags: [game]
      summary: Get NPCs encountered in this session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: NPCs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Npc"
  /sessions/{sessionId}/journal:
    get:
      operationId: getSessionJournal
      tags: [game]
      summary: Get campaign journal entries for this session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
  /sessions/{sessionId}/map:
    get:
      operationId: getSessionMap
      tags: [game]
      summary: Get world map for this session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: World map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorldMap"
  /sessions/{sessionId}/combat:
    get:
      operationId: getCombatState
      tags: [game]
      summary: Get current combat state
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Combat state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CombatState"
    put:
      operationId: updateCombatState
      tags: [game]
      summary: Update combat state
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCombatRequest"
      responses:
        "200":
          description: Updated combat state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CombatState"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    PartyMember:
      type: object
      properties:
        playerId:
          type: integer
        username:
          type: string
        characterName:
          type: string
        race:
          type: string
        class:
          type: string
        level:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        xp:
          type: integer
        isDead:
          type: boolean
      required: [playerId, username, characterName, race, class, level, hp, maxHp, xp]
    DmInjectRequest:
      type: object
      properties:
        creatorId:
          type: integer
        event:
          type: string
      required: [creatorId, event]
...
    Npc:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        name:
          type: string
        description:
          type: string
        disposition:
          type: string
        firstMet:
          type: string
      required: [id, sessionId, name, description, disposition]
    JournalEntry:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        summary:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, sessionId, summary, createdAt]
    WorldMap:
      type: object
      properties:
        sessionId:
          type: integer
        locations:
          type: array
          items:
            type: string
        currentLocation:
          type: string
        ascii:
          type: string
      required: [sessionId, locations]
    CombatState:
      type: object
      properties:
        sessionId:
          type: integer
        active:
          type: boolean
        round:
          type: integer
        combatants:
          type: array
          items:
            $ref: "#/components/schemas/Combatant"
      required: [sessionId, active, round, combatants]
    Combatant:
      type: object
      properties:
        name:
          type: string
        hp:
          type: integer
        maxHp:
          type: integer
        initiative:
          type: integer
        isPlayer:
          type: boolean
      required: [name, hp, maxHp, initiative, isPlayer]
    UpdateCombatRequest:
      type: object
      properties:
        active:
          type: boolean
        round:
          type: integer
        combatants:
          type: array
          items:
            $ref: "#/components/schemas/Combatant"
      required: [active, round, combatants]
...
    UpdateMemberStatsRequest:
      type: object
      properties:
        hp:
          type: integer
        maxHp:
          type: integer
        level:
          type: integer
        xp:
          type: integer
      required: [hp, maxHp, level, xp]
...